+++ packages/terra-framework-docs/src/terra-dev-site/doc/popup/example/PopupStandard.jsx
@@ -16,15 +16,24 @@ function PopupStandard() {
   const getButtonNode = () => buttonElement.current;
 
   const handleButtonClick = () => {
+    if (!open) {
+      document.querySelector('#root').setAttribute('inert', 'true');
+    }
     setOpen(true);
   };
 
   const handleRequestClose = () => {
     setOpen(false);
+    document.querySelector('#root').removeAttribute('inert');
+    buttonElement.current.focus();
   };
 
   return (
     <React.Fragment>
+      <div>
+        To prevent focus from shifting with up/down arrow keys when virtual cursor mode is on in Jaws use inert attribute as shown in the example.
+        Using inert makes the content outside the popup inaccessible to screen readers and keyboard interactions while the popup is active, ensuring keyboard focus is trapped within the popup for better accessibility.
+      </div>
       <button type="button" className={cx('popup-button')} onClick={handleButtonClick} ref={setButtonNode}>Medications</button>
       <Popup
         isOpen={open}